refactor(goals): extract useYearMonthlySaving hook (step 11)
Move budget-year savings calculation into a reusable hook at
src/pages/goal/hooks/useYearMonthlySaving.ts. GoalDetail drops
~40 lines of inline budget-parsing logic.
